Refactoring of modules
